Grammatical/autorespond
Changed your to you are in database.py Didn't know if apostrophes would break anything. Updated events.py with responses for the ShadowApples guy that people continue to come to this Discord for. Also bastardized something together to respond to the words slow and drive, but I don't know if I quite nailed it. Might help funnel people with additional storage issues to support in the interim while things are being worked on.